@@ -20,7 +20,8 @@
     "bootstrap": "^5.3.2",
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.19",
-    "next": "^13.4.19",
+    "js-cookie": "^3.0.5",
+    "next": "^13.5.6",
     "oracledb": "^6.1.0",
     "postcss": "^8.4.30",
     "postcss-cli": "^10.1.0",
@@ -30,9 +31,11 @@
     "react-dom": "^18.2.0",
     "react-markdown": "^9.0.0",
     "react-modal": "^3.16.1",
+    "router": "^1.3.8",
     "tailwindcss": "3.3.3"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.5",
     "@types/node": "^20.6.2",
     "@types/react-modal": "^3.16.1",
     "prisma": "^5.3.1",
