@@ -1,30 +1,25 @@
 'use client';
 
-import React, { useState, useEffect } from 'react';
-
-import useFetch from '../../hooks/useFetch';
-import './state.css';
 import axios from 'axios';
+import React, { useState, useEffect } from 'react';
 import useMongoNickname from '../../hooks/mongoNickname';
+import './state.css';
 
-interface IMemberstate {
+interface IMember {
   state: boolean;
   createDate: string;
   passwordChangeDate: string;
   email: string;
-  nickname: string;
   password: string;
+  nickname: string;
   phoneNum: string;
   address: string;
   warning: number;
 }
-const Member: React.FC<{ member: IMemberstate }> = ({ member }) => {
-  const {
-    nickname: mongoNickname,
-    loading,
-    error,
-  } = useMongoNickname(member.email);
-  console.log('매매맴버버버버:', Member);
+
+const Member: React.FC<{ member: IMember }> = ({ member }) => {
+  const { nickname: mongoNickname, loading } = useMongoNickname(member.email);
+
   return (
     <tr>
       <td className={member.state ? 'statusNormal' : 'statusSuspended'}>
@@ -41,21 +36,14 @@ const Member: React.FC<{ member: IMemberstate }> = ({ member }) => {
     </tr>
   );
 };
-console.log('매매맴버버버버:', Member);
 
-// Search 컴포넌트 정의
 export default function Search() {
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === 'Enter') {
-      handleSearch();
-    }
-  };
-
   const [searchCategory, setSearchCategory] = useState('email');
-  const [searchEmail, setSearchEmail] = useState(''); // Step 1
-  const [searchedMembers, setSearchedMembers] = useState<IMemberstate[]>([]); // 새로운 상태 추가
+  const [searchEmail, setSearchEmail] = useState('');
+  const [members, setMembers] = useState<IMember[]>([]);
+  const [searchedMembers, setSearchedMembers] = useState<IMember[]>([]);
   const [searchClicked, setSearchClicked] = useState(false);
-  // Step 2
+
   const handleSearch = async (): Promise<void> => {
     setSearchClicked(true);
     try {
@@ -67,60 +55,40 @@ export default function Search() {
       const searchData =
         searchCategory === 'email'
           ? { email: searchEmail }
-          : { phoneNum: searchEmail }; // 이 부분은 email 상태 변수명을 더 일반적인 이름으로 변경하는 것이 좋겠습니다.
+          : { phoneNum: searchEmail };
 
       const response = await axios.post(searchUrl, searchData);
-      const filteredSearchResults = response.data.filter(
-        (member: IMemberstate) => !member.state
-      );
-      setSearchedMembers(filteredSearchResults);
+      setSearchedMembers(response.data.filter((m: IMember) => !m.state));
     } catch (error) {
       console.error('Search failed', error);
     }
   };
 
-  // useFetch로 백엔드에서 멤버 리스트 가져옴
-  const [members, setMembers] = useState<IMemberstate[]>([]);
+  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
+    if (event.key === 'Enter') {
+      handleSearch();
+    }
+  };
 
   useEffect(() => {
-    async function fetchProfileAndCombine() {
+    async function fetchMembers() {
       try {
-        const responseProfile = await fetch(`/api/profileList`);
-        const dataProfile = await responseProfile.json();
-
-        const responseMembers = await fetch(
-          'http://127.0.0.1:8080/member/memberState'
-        );
-        const dataMembers = await responseMembers.json();
-
-        const combinedData = dataMembers.map((member: IMemberstate) => {
-          const matchingProfile = dataProfile.find(
-            (profile: any) => profile.email === member.email
-          );
-          return {
-            ...member,
-            nickname: matchingProfile ? matchingProfile.nickname : null,
-          };
-        });
-        console.log('데이터:', combinedData);
-
-        setMembers(combinedData);
+        const response = await fetch('http://127.0.0.1:8080/admin/memberList');
+        const data = await response.json();
+        setMembers(data.filter((m: IMember) => !m.state));
       } catch (error) {
         console.error('Error fetching data:', error);
       }
     }
-
-    fetchProfileAndCombine();
+    fetchMembers();
   }, []);
-
-  // 경고 수 증가 처리를 위한 핸들러 함수
 
   if (!members) return <span>Loading...</span>;
 
   return (
     <>
       {members.length === 0 ? (
-        <span>악성 유저 정보 없음</span>
+        <span>유저정보 없음</span>
       ) : (
         <div className='table-container'>
           <table className='table'>
@@ -129,15 +97,11 @@ export default function Search() {
                 <th
                   colSpan={9}
                   style={{ textAlign: 'center', fontSize: '30px' }}>
-                  정지 유저 정보
+                  정지된 회원 정보
                 </th>
               </tr>
               <tr>
-                <th
-                  colSpan={13}
-                  style={{
-                    textAlign: 'center',
-                  }}>
+                <th colSpan={9} style={{ textAlign: 'center' }}>
                   <select
                     className='search-input1'
                     value={searchCategory}
